gunyah: Fix possible overflow in the API gh_provide_user_mem
Assign the u32 to u16 which probably overflow when
it exceeds u16 limit. Fix it by change n_sgl_entries
type from u32 to unsigned long to avoid n_sgl_entries
is overflow when there are so many pages needed to do
hyp assgin. And check the n_sgl_entries is larger
than U16_MAX or not before assign to u16.

+++ drivers/virt/gunyah/gh_private.h
@@ -20,7 +20,7 @@ struct gh_vm_user_mem {
 	struct sg_table *sgt;
 	struct page **pages;
 	unsigned long npages;
-	u32 n_sgl_entries;
+	unsigned long n_sgl_entries;
 	struct gh_sgl_entry *sgl_entries;
 	gh_memparcel_handle_t mem_handle;
 };
